services/ai-agent/pyproject.toml
[tool.poetry]
name = "qgen-ai-agent"
version = "0.1.0"
description = "AI Research Assistant for Questionnaire & Banner Planning"
authors = ["Your Team <team@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
python-dotenv = "^1.0.0"

# AI & LLM
openai = "^1.3.0"
anthropic = "^0.7.0"
langchain = "^0.1.0"
langchain-openai = "^0.0.2"

# Vector DB & Embeddings
pgvector = "^0.2.3"
sentence-transformers = "^2.2.2"

# Database
supabase = "^2.0.0"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.23"

# Utilities
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
httpx = "^0.25.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
ruff = "^0.1.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
